# specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCategoryResponseDto"
                }
              }
            }
          }
        },
        "summary": "List all categories with their subcategories",
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "Get a category by ID (includes subcategories)",
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}/subcategories": {
      "get": {
        "operationId": "CategoriesController_findSubCategories",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSubCategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "List subcategories for a category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/admin/categories": {
      "get": {
        "operationId": "AdminCategoriesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCategoryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List all categories",
        "tags": [
          "Admin – Categories"
        ]
      },
      "post": {
        "operationId": "AdminCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a category",
        "tags": [
          "Admin – Categories"
        ]
      }
    },
    "/admin/categories/{id}": {
      "patch": {
        "operationId": "AdminCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update a category",
        "tags": [
          "Admin – Categories"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete a category",
        "tags": [
          "Admin – Categories"
        ]
      }
    },
    "/admin/categories/{id}/subcategories": {
      "post": {
        "operationId": "AdminCategoriesController_createSubCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubCategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Add a subcategory to a category",
        "tags": [
          "Admin – Categories"
        ]
      }
    },
    "/admin/categories/{id}/subcategories/{subId}": {
      "patch": {
        "operationId": "AdminCategoriesController_updateSubCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "description": "Subcategory UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubCategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Category or subcategory not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update a subcategory",
        "tags": [
          "Admin – Categories"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_removeSubCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "description": "Subcategory UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Subcategory deleted"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Category or subcategory not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete a subcategory",
        "tags": [
          "Admin – Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2Shop API",
    "description": "REST API for the O2Shop e-commerce platform. Public endpoints need no authentication. Customer endpoints require a JWT bearer token (from POST /auth/login). Admin endpoints additionally require the `admin` role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access_token value from POST /auth/login."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "example": [
              "email must be an email"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "SubCategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "t-shirts"
          },
          "displayName": {
            "type": "string",
            "example": "T-Shirts"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id",
          "slug",
          "displayName",
          "categoryId"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "clothing"
          },
          "displayName": {
            "type": "string",
            "example": "Clothing"
          },
          "subCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubCategoryResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "displayName",
          "subCategories",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedCategoryResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponseDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ]
      },
      "PaginatedSubCategoryResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubCategoryResponseDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "clothing",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "displayName": {
            "type": "string",
            "example": "Clothing"
          }
        },
        "required": [
          "slug",
          "displayName"
        ]
      },
      "CreateSubCategoryDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "t-shirts",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "displayName": {
            "type": "string",
            "example": "T-Shirts"
          }
        },
        "required": [
          "slug",
          "displayName"
        ]
      },
...
    }
  }
}